requirements.txt
streamlit
tensorflow
transformers
torch
numpy
pandas
scikit-learn
pillow
pyngrok
sqlite3
